lesson08 cart minor bugs removed

+++ lesson08/controllers/config.php
@@ -9,15 +9,12 @@ const PASS = "";
 mysqli_report(MYSQLI_REPORT_ERROR | MYSQLI_REPORT_STRICT);
 $connect = mysqli_connect(SERVER,LOGIN,PASS,DB);
 
-// Variables
-
-$goods = goodsAll($connect);
-$goodsInCart = getGoodsInCart($connect);
-
 // Functions
 
-function getGoodsInCart($connect) {
-    $sqlQuery = "SELECT goodId, title, price * count AS sum, count FROM goods INNER JOIN cart ON cart.goodId = goods.id AND userId = {$_SESSION['userId']}";
+function getGoodsInCart($connect, $status=0) {
+    $sqlQuery = "SELECT goodId, title, price * count AS sum, count FROM goods
+    INNER JOIN cart ON cart.goodId = goods.id
+    AND userId = {$_SESSION['userId']} AND status = {$status}";
     $answer = mysqli_query($connect, $sqlQuery);
     if(!$answer) {
         die(mysqli_error($connect));
@@ -33,7 +30,7 @@ function cartManager($connect, $goodId, $action) {
     $message = "";
     switch ($action) {
         case 'addGood':
-            $sqlQuery = "select id from cart where goodId = ". $goodId;
+            $sqlQuery = "select id from cart where goodId = $goodId and userId = {$_SESSION['userId']} and status = 0";
             $answer = mysqli_query($connect, $sqlQuery);
             $goodInCart = mysqli_fetch_assoc($answer)['id'];
             if($goodInCart == null) {
@@ -45,19 +42,19 @@ function cartManager($connect, $goodId, $action) {
             }
             break;
         case 'delGood':
-            $sqlQuery = "delete from cart where goodId = $goodId";
+            $sqlQuery = "delete from cart where goodId = $goodId and userId = {$_SESSION['userId']}";
             break;
         case 'addCount':
-            $sqlQuery = "update cart set count=count+1 where goodId = $goodId";
+            $sqlQuery = "update cart set count=count+1 where goodId = $goodId and userId = {$_SESSION['userId']}";
             break;
         case 'remCount':
-            $sqlQuery = "select count from cart where goodId = $goodId";
+            $sqlQuery = "select count from cart where goodId = $goodId and userId = {$_SESSION['userId']}";
             $answer = mysqli_query($connect, $sqlQuery);
             $count = mysqli_fetch_assoc($answer)['count'];
             if($count > 1){   
-                $sqlQuery = "update cart set count=count-1 where goodId = $goodId";    
+                $sqlQuery = "update cart set count=count-1 where goodId = $goodId and userId = {$_SESSION['userId']}";    
             } else {
-                $sqlQuery = "delete from cart where goodId = $goodId";         
+                $sqlQuery = "delete from cart where goodId = $goodId and userId = {$_SESSION['userId']}";         
             }
             break;
     }
